+++ packages/block-library/src/quote/test/edit.native.js
@@ -10,7 +10,7 @@ import {
 	fireEvent,
 	within,
 	waitFor,
-	changeAndSelectTextOfRichText,
+	typeInRichText,
 } from 'test/helpers';
 
 /**
@@ -49,11 +49,7 @@ describe( 'Quote', () => {
 		// screen.debug();
 		let quoteTextInput =
 			within( quoteBlock ).getByPlaceholderText( 'Start writing…' );
-		const string = 'A great statement.';
-		changeAndSelectTextOfRichText( quoteTextInput, string, {
-			selectionStart: string.length,
-			selectionEnd: string.length,
-		} );
+		typeInRichText( quoteTextInput, 'A great statement.' );
 		fireEvent( quoteTextInput, 'onKeyDown', {
 			nativeEvent: {},
 			preventDefault() {},
@@ -63,12 +59,12 @@ describe( 'Quote', () => {
 			within( quoteBlock ).getAllByPlaceholderText(
 				'Start writing…'
 			)[ 1 ];
-		changeAndSelectTextOfRichText( quoteTextInput, 'Again.' );
+		typeInRichText( quoteTextInput, 'Again.' );
 		const citationTextInput =
 			within( citationBlock ).getByPlaceholderText( 'Add citation' );
-		changeAndSelectTextOfRichText( citationTextInput, 'A person', {
-			selectionStart: 2,
-			selectionEnd: 2,
+		typeInRichText( citationTextInput, 'A person', {
+			finalSelectionStart: 2,
+			finalSelectionEnd: 2,
 		} );
 		fireEvent( citationTextInput, 'onKeyDown', {
 			nativeEvent: {},