+++ packages/edit-site/src/hooks/commands/use-wp-admin-commands.js
@@ -6,6 +6,7 @@ import { __, sprintf } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
 import { addQueryArgs } from '@wordpress/url';
 import { useMemo } from '@wordpress/element';
+import { plus } from '@wordpress/icons';
 
 /**
  * Internal dependencies
@@ -43,8 +44,9 @@ const getWPAdminAddCommandLoader = ( postType ) =>
 		const commands = useMemo(
 			() => [
 				{
-					name: 'core/wp-admin/add-' + postType,
+					name: 'core/wp-admin/add-' + postType + '-' + search,
 					label,
+					icon: plus,
 					callback: () => {
 						document.location.href = addQueryArgs( newPostLink, {
 							post_type: postType,