+++ packages/env/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   Use test environment's `WP_SITEURL` instead of `WP_TESTS_DOMAIN` as the WordPress URL.
 -   Automatically add the environment's port to `WP_TESTS_DOMAIN`.
+-   `run` command now has a `--env-cwd` option to set the working directory in the container for the command to execute from.
 
 ## 5.16.0 (2023-04-12)